+++ rtk-crud-cromos/src/components/StickersListHeader.js
@@ -1,6 +1,5 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { Link } from 'react-router-dom'
-import StickerCard from './StickerCard'
 
 function StickersListHeader() {
     const stickers = useSelector(state => state.stickers)
